move iterator process from Interactor to UI
