+++ TrainDelayReporter_iOS/UI/Components/HamburgerMenu.swift
@@ -12,7 +12,19 @@ import Combine
 
 struct HamburgerMenu: View {
     
-    @ObservedObject var regionListInteractor = RegionListInteractor()
+    //@ObservedObject var regionListInteractor = RegionListInteractor()
+    @State var isComplete:Bool = false
+    @State var regionStructs:[Region] = []
+    
+    let regions:[String: Int] = [
+        "北海道": 0x02C03C,
+        "東日本": 0x37863F,
+        "西日本": 0x0072BA,
+        "東海": 0xFF7E1C,
+        "四国": 0x00ACD1,
+        "九州": 0xF62D36,
+        "私鉄": 0x98A9D6
+    ]
     
     var body: some View {
         GeometryReader { geometry in
@@ -20,10 +32,10 @@ struct HamburgerMenu: View {
                 Text("地域を選択")
                     .font(.system(size: 24))
                 Divider()
-                if self.regionListInteractor.isComplete == false {
+                if self.isComplete == false {
                     Text("地域情報を取得中")
                 }
-                QGrid(self.regionListInteractor.regionStructs,
+                QGrid(self.regionStructs,
                       columns: 1,
                       vSpacing: 20,
                       hSpacing: 20,
@@ -35,26 +47,19 @@ struct HamburgerMenu: View {
             }
             .padding(.horizontal, 20)
             .onAppear(perform: {
-                self.regionListInteractor.setRegions()
+                self.setRegions()
             })
         }
     }
-}
-
-// [参考](https://qiita.com/masa7351/items/e634f50ade5915f13397)
-extension Color {
-    init(hex: Int, alpha: Double = 1) {
-        let components = (
-            R: Double((hex >> 16) & 0xff) / 255,
-            G: Double((hex >> 08) & 0xff) / 255,
-            B: Double((hex >> 00) & 0xff) / 255
-        )
-        self.init(
-            .sRGB,
-            red: components.R,
-            green: components.G,
-            blue: components.B,
-            opacity: alpha
-        )
+    
+    /// Yield Region cells for HamburgerMenu
+    public func setRegions() {
+        for region in regions {
+            self.regionStructs.append(Region(name: region.key, colorCode: region.value))
+        }
+        // Exec status change in main thred to avoid an error
+        DispatchQueue.main.async {
+            self.isComplete = true
+        }
     }
 }
